fix: resolve circular dependencies
refs #109
